Added stripe
Refactored timezone handling
Refactored tokens handling

+++ Services/StateHandlers/ContactInfoHandler.cs
@@ -1,5 +1,5 @@
 using Microsoft.EntityFrameworkCore;
-using Telegram.Bot.Services.Constants;
+using Telegram.Bot.Commands.Common;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.ReplyMarkups;
 
@@ -7,6 +7,7 @@ namespace Telegram.Bot.Services.StateHandlers;
 
 public class ContactInfoHandler : BaseStateHandler
 {
+    private readonly IMainMenuCommandHandler _mainMenuCommandHandler;
     public override List<string> StateNames => new List<string> {  "WaitingForContactInfo" };
 
     public ContactInfoHandler(
@@ -15,38 +16,47 @@ public class ContactInfoHandler : BaseStateHandler
         ILogger<ContactInfoHandler> logger,
         BookingDbContext dbContext,
         ICompanyCreationStateService companyCreationStateService,
-        ITranslationService translationService)
+        ITranslationService translationService,
+        IMainMenuCommandHandler mainMenuCommandHandler)
         : base(botClient, userStateService, logger, dbContext, companyCreationStateService, translationService)
     {
+        _mainMenuCommandHandler = mainMenuCommandHandler;
     }
 
     public override async Task HandleAsync(long chatId, string state, Message message, CancellationToken cancellationToken)
     {
-        if (string.IsNullOrEmpty(message.Text))
+        var language = await UserStateService.GetLanguageAsync(chatId, cancellationToken);
+        if (message.Contact != null)
+        {
+            var phoneNumber = message.Contact.PhoneNumber;
+            var client = await DbContext.Clients.FirstOrDefaultAsync(c => c.ChatId == chatId);
+            if (client == null)
+            {
+                await BotClient.SendMessage(
+                    chatId: message.Chat.Id,
+                    text: TranslationService.Get(language, "NoClientFound"), 
+                    replyMarkup: new ReplyKeyboardRemove(),
+                    cancellationToken: cancellationToken);
+                return;
+            }
+
+            client.PhoneNumber = phoneNumber;
+            DbContext.Clients.Update(client);
+            await DbContext.SaveChangesAsync(cancellationToken);
+            UserStateService.RemoveConversation(chatId);
+            await BotClient.SendMessage(
+                chatId: message.Chat.Id,
+                text: TranslationService.Get(language, "DataSaved"), 
+                replyMarkup: new ReplyKeyboardRemove(),
+                cancellationToken: cancellationToken);
+            await _mainMenuCommandHandler.ShowClientMainMenuAsync(chatId, language, cancellationToken);
+        } 
+        else
         {
             await BotClient.SendMessage(
                 chatId: chatId,
-                text: TranslationService.Get(await UserStateService.GetLanguageAsync(chatId, cancellationToken), "ContactInfoRequired"),
+                text: TranslationService.Get(language, "ContactInfoRequired"),
                 cancellationToken: cancellationToken);
-            return;
         }
-
-        var adminChatId = await DbContext.Companies
-            .Where(c => c.Alias == "demo")
-            .Select(c => c.Token.ChatId)
-            .FirstOrDefaultAsync();
-            
-        await BotClient.SendMessage(
-            chatId: adminChatId,
-            text: $"📩 New company creation request!\n\nFrom user ID: {chatId}\nContact: {message.Text}",
-            cancellationToken: cancellationToken);
-
-        await BotClient.SendMessage(
-            chatId: chatId,
-            text: TranslationService.Get(await UserStateService.GetLanguageAsync(chatId, cancellationToken), "NewContactThanks"),
-            replyMarkup: new ReplyKeyboardRemove(),
-            cancellationToken: cancellationToken);
-
-        UserStateService.RemoveConversation(chatId);
     }
-} 
+}
